# src/docs/swagger/collections.yaml
openapi: 3.0.0
info:
  title: Shopease
  description: The API collection for Shopease
  version: 1.0.0
servers:
  - url: http://localhost:9000/api/v1
    description: Shopease Development Server
  - url: http://api.shopease.shubhampurwar.in/api/v1
    description: Shopease Production Server
...
paths:
...
  /auth/password/forgot:
    post:
      tags:
        - Authentication
      description: Sends an email to the user with reset password link
      operationId: forgotPassword
      requestBody:
        description: A JSON object containing an email address
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Password reset email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /auth/password/reset/{token}:
    put:
      tags:
        - Authentication
      description: Allows users to reset their account password
      operationId: resetPassword
      parameters:
        - in: path
          name: token
          required: true
          description: Reset password token
          schema:
            type: string
      requestBody:
        description: A JSON object containing new and confirmed passwords
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                confirmPassword:
                  type: string
      responses:
        '200':
          description: Password reset success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
...
  /addresses:
    get:
      tags:
        - Addresses
      description: Allows users to fetch their addresses
      operationId: getAddresses
      responses:
        '200':
          description: A JSON object containing a list of addresses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Address'

    post:
      tags:
        - Addresses
      description: Allows users to add a new address
      operationId: addNewAddress
      requestBody:
        description: A JSON object containing address info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressBody'
      responses:
        '201':
          description: A JSON object containing a new address
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Address'
  /addresses/{addressId}:
    get:
      tags:
        - Addresses
      description: Allows users to fetch one of their addresses by ID
      operationId: getAddressById
      parameters:
        - in: path
          name: addressId
          required: true
          description: ID of the address to be fetched
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an address
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Address'

    put:
      tags:
        - Addresses
      description: Allows users to update their address
      operationId: updateAddress
      parameters:
        - in: path
          name: addressId
          required: true
          description: ID of the address to be updated
          schema:
            type: string
      requestBody:
        description: A JSON object containing new address info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressBody'
      responses:
        '200':
          description: A JSON object containing an updated address
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Address'

    delete:
      tags:
        - Addresses
      description: Allows users to delete their address
      operationId: deleteAddress
      parameters:
        - in: path
          name: addressId
          required: true
          description: ID of the address to be deleted
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing ID of the deleted address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
  /addresses/{addressId}/default:
    put:
      tags:
        - Addresses
      description: Allows users to set one of their addresses as the default
      operationId: setDefaultAddress
      parameters:
        - in: path
          name: addressId
          required: true
          description: ID of the address to be set as the default
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing a default address
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Address'
...
  /brands/{brandId}:
    get:
      tags:
        - Brands
      description: Fetches a brand by ID
      operationId: getBrandById
      parameters:
        - in: path
          name: brandId
          required: true
          description: ID of the brand to be fetched
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing a brand
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Brand'
  /admin/brands:
    post:
      tags:
        - Brands
      description: Allows admins to add a new brand
      operationId: addNewBrand
      requestBody:
        description: Payload containing name and logo of a brand
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrandBody'
            encoding:
              logo:
                contentType: image/*
      responses:
        '201':
          description: A JSON object containing a new brand
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Brand'
  /admin/brands/{brandId}:
    post:
      tags:
        - Brands
      description: Allows admins to update a brand
      operationId: updateBrand
      parameters:
        - in: path
          name: brandId
          required: true
          description: ID of the brand to be updated
          schema:
            type: string
      requestBody:
        description: Payload containing name and logo of a brand
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrandBody'
            encoding:
              logo:
                contentType: image/*
      responses:
        '200':
          description: A JSON object containing an updated brand
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Brand'
...
components:
...
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ID:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
...
    Address:
      type: object
      properties:
        id:
          type: string
        fullname:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        landmark:
          type: string
        country:
          type: string
        state:
          type: string
        city:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
        user:
          type: string
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AddressBody:
      type: object
      properties:
        fullname:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        landmark:
          type: string
        country:
          type: string
        state:
          type: string
        city:
          type: string
        postalCode:
          type: string
        isDefault:
          type: boolean
          default: false
...
    Brand:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logo:
          type: object
          properties:
            url:
              type: string
            publicId:
              type: string
        createdBy:
          type: string
        lastUpdatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BrandBody:
      type: object
      properties:
        logo:
          type: string
          format: binary
        name:
          type: string
      required:
        - logo
        - name
...